always use single threaded tests
The binary currently performs all operations based on the current
working directory. This made sense as I was doing the initial spike, but
complicates testing because current working directory is process local,
not thread local. This means that multithreaded integration tests that
use file system fixtures and change directories are unpredictable and
error prone.

While it may not be ideal to rely on the processes current working
directory, I'd rather add some tests then perform a refactor first.