add not enough commits test
This is an integration test that checks for the output of running `teva`
without enough commits between `main` and `HEAD` on the feature branch.

Instead of dynamically creating the test scaffolding, this instead adds
a git sub-module with real files and state. Of all of the setup
strategies I explored, this was by far the most debug-able and simplest
to maintain.